requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
asn1crypto=1.4.0=py_0
blas=1.0=mkl
brotlipy=0.7.0=py37h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.1.19=h06a4308_1
certifi=2020.12.5=py37h06a4308_0
cffi=1.14.0=py37h2e261b9_0
chardet=4.0.0=py37h06a4308_1003
cryptography=2.3.1=py37hc365091_0
cudatoolkit=10.2.89=hfd86e86_1
ffmpeg=4.3=hf484d3e_0
freetype=2.10.4=h5ab3b9f_0
gmp=6.2.1=h2531618_2
gnutls=3.6.5=h71b1129_1002
idna=2.10=pyhd3eb1b0_0
intel-openmp=2020.2=254
jpeg=9b=h024ee3a_2
lame=3.100=h7b6447c_0
lcms2=2.11=h396b838_0
libedit=3.1.20191231=h14c3975_1
libffi=3.2.1=hf484d3e_1007
libgcc-ng=9.1.0=hdf63c60_0
libiconv=1.15=h63c8f33_5
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuv=1.40.0=h7b6447c_0
lz4-c=1.9.3=h2531618_0
mkl=2020.2=256
mkl-service=2.3.0=py37he8ac12f_0
mkl_fft=1.3.0=py37h54f3939_0
mkl_random=1.1.1=py37h0573a6f_0
ncurses=6.2=he6710b0_1
nettle=3.4.1=hbb512f6_0
ninja=1.10.2=py37hff7bd54_0
numpy=1.19.2=py37h54aff64_0
numpy-base=1.19.2=py37hfa32c7d_0
olefile=0.46=py_0
openh264=2.1.0=hd408876_0
openssl=1.0.2u=h7b6447c_0
pillow=8.1.2=py37he98fc37_0
pip=21.0.1=py37h06a4308_0
pycparser=2.20=py_2
pyopenssl=19.0.0=py37_0
pysocks=1.7.1=py37_1
python=3.7.0=h6e4f718_3
pytorch=1.8.0=py3.7_cuda10.2_cudnn7.6.5_0
readline=7.0=h7b6447c_5
requests=2.25.1=pyhd3eb1b0_0
setuptools=52.0.0=py37h06a4308_0
six=1.15.0=pyhd3eb1b0_0
sqlite=3.33.0=h62c20be_0
tk=8.6.10=hbc83047_0
torchaudio=0.8.0=py37
torchtext=0.9.0=py37
torchvision=0.9.0=py37_cu102
tqdm=4.56.0=pyhd3eb1b0_0
typing_extensions=3.7.4.3=pyha847dfd_0
urllib3=1.26.3=pyhd3eb1b0_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
